+++ backend/backends/backends.go
@@ -1,7 +1,8 @@
 package backends
 
 import (
-	_ "github.com/jtblin/gostatsd/backend/backends/datadog"  // imports backends to avoid cycle error
-	_ "github.com/jtblin/gostatsd/backend/backends/graphite" // imports backends to avoid cycle error
-	_ "github.com/jtblin/gostatsd/backend/backends/stdout"   // imports backends to avoid cycle error
+	_ "github.com/jtblin/gostatsd/backend/backends/datadog"     // imports backends to avoid cycle error
+	_ "github.com/jtblin/gostatsd/backend/backends/graphite"    // imports backends to avoid cycle error
+	_ "github.com/jtblin/gostatsd/backend/backends/statsdaemon" // imports backends to avoid cycle error
+	_ "github.com/jtblin/gostatsd/backend/backends/stdout"      // imports backends to avoid cycle error
 )